+++ lib/pages/create/create_measurement_dialog.dart
@@ -155,7 +155,7 @@ class _MeasurementDialogState extends State<MeasurementDialog> {
                     Text(
                       dataType.displayName,
                       style: const TextStyle(
-                        color: Colors.black,
+                        color: habitCardTextColor,
                         fontSize: fontSizeMedium,
                       ),
                     ),
@@ -184,7 +184,7 @@ class _MeasurementDialogState extends State<MeasurementDialog> {
                       DateTimeField(
                         dateTime: measurementTime,
                         labelText: localizations.addMeasurementDateLabel,
-                        style: newInputStyle().copyWith(color: Colors.black),
+                        style: dialogInputStyle(),
                         setDateTime: (picked) {
                           setState(() {
                             measurementTime = picked;
@@ -198,9 +198,10 @@ class _MeasurementDialogState extends State<MeasurementDialog> {
                         decoration: createDialogInputDecoration(
                           labelText: '${dataType.displayName} '
                               '${dataType.unitName.isNotEmpty ? '[${dataType.unitName}] ' : ''}',
+                          style: dialogInputStyle(),
                         ),
                         keyboardAppearance: keyboardAppearance(),
-                        style: newInputStyle().copyWith(color: Colors.black),
+                        style: dialogInputStyle(),
                         validator: numericValidator(),
                         name: 'value',
                         keyboardType: const TextInputType.numberWithOptions(
@@ -213,9 +214,10 @@ class _MeasurementDialogState extends State<MeasurementDialog> {
                         key: const Key('measurement_comment_field'),
                         decoration: createDialogInputDecoration(
                           labelText: localizations.addMeasurementCommentLabel,
+                          style: dialogInputStyle(),
                         ),
                         keyboardAppearance: keyboardAppearance(),
-                        style: newInputStyle().copyWith(color: Colors.black),
+                        style: dialogInputStyle(),
                         name: 'comment',
                       ),
                       inputSpacer,