+++ lib/themes/theme.dart
@@ -23,6 +23,8 @@ class AppTheme {
 const double chipBorderRadius = 8;
 const mainFont = 'PlusJakartaSans';
 
+const habitCardTextColor = Colors.black87;
+
 final inputBorder = OutlineInputBorder(
   borderRadius: BorderRadius.circular(chipBorderRadius),
   borderSide: BorderSide(color: styleConfig().secondaryTextColor),
@@ -76,8 +78,8 @@ TextStyle inputStyle() => TextStyle(
       fontSize: fontSizeMedium,
     );
 
-TextStyle newInputStyle() => TextStyle(
-      color: styleConfig().primaryTextColor,
+TextStyle dialogInputStyle() => const TextStyle(
+      color: habitCardTextColor,
       fontSize: fontSizeMedium,
     );